1.修改系统程序，按照对接的方式发送SOE与COS事件； 2.三线俩表法计算数据，电气连接必须是双PT采用V型连接，电流输入为A、C、零序电流； 3.软遥信重新进行编号调整，根据对接进行调整；

+++ RN2012-F30-DSP/dsp/app/fft.h
@@ -54,7 +54,7 @@ Void FFT_Task(UArg arg0, UArg arg1);
 Void GetTempture(CurrentPaStr *remotevalue);
 float P_Value(UInt8 u1_ch, UInt8 u2_ch);
 float Q_Value(UInt8 u1_ch, UInt8 u2_ch);
-float SumDC(float *data, Int8 len);
+float SumDC(float *data, UInt16 len);
 
 //Void FFT(ANSMP_Stru *Input, BaseValueType *Out, UChar flagtmp, float *adjust);
